move app lifecycle reactor into a widget
