+++ lib/trufi_app.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:package_info/package_info.dart';
 import 'package:trufi_core/blocs/app_review_cubit.dart';
 import 'package:trufi_core/blocs/home_page_cubit.dart';
 import 'package:trufi_core/blocs/locations/favorite_locations_cubit/favorite_locations_cubit.dart';
@@ -16,7 +15,6 @@ import 'package:trufi_core/pages/home/home_page.dart';
 import 'package:trufi_core/repository/shared_preferences_repository.dart';
 import 'package:trufi_core/trufi_configuration.dart';
 import 'package:trufi_core/trufi_observer.dart';
-import 'package:trufi_core/widgets/app_review_dialog.dart';
 import 'package:uuid/uuid.dart';
 
 import './blocs/bloc_provider.dart';
@@ -29,6 +27,7 @@ import './pages/saved_places.dart';
 import './pages/team.dart';
 import './widgets/trufi_drawer.dart';
 import 'blocs/locations/saved_places_locations_cubit/saved_places_locations_cubit.dart';
+import 'pages/app_lifecycle_reactor.dart';
 import 'services/plan_request/online_repository.dart';
 import 'services/search_location/offline_search_location.dart';
 
@@ -158,56 +157,9 @@ class TrufiApp extends StatelessWidget {
   }
 }
 
-class AppLifecycleReactor extends StatefulWidget {
-  const AppLifecycleReactor({
-    Key key,
-    @required this.child,
-  }) : super(key: key);
 
-  final Widget child;
 
-  @override
-  _AppLifecycleReactorState createState() => _AppLifecycleReactorState();
-}
-
-class _AppLifecycleReactorState extends State<AppLifecycleReactor>
-    with WidgetsBindingObserver {
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addObserver(this);
-  }
-
-  @override
-  void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
-    super.dispose();
-  }
-
-  @override
-  Future<void> didChangeAppLifecycleState(AppLifecycleState state) async {
-    final locationProviderCubit = context.read<LocationProviderCubit>();
-
-    if (state == AppLifecycleState.resumed) {
-      final appReviewBloc = BlocProvider.of<AppReviewCubit>(context);
-      final packageInfo = await PackageInfo.fromPlatform();
-      if (await appReviewBloc.isAppReviewAppropriate(packageInfo)) {
-        showAppReviewDialog(context);
-        appReviewBloc.markReviewRequestedForCurrentVersion(packageInfo);
-      }
-      locationProviderCubit.start();
-    } else {
-      locationProviderCubit.stop();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return widget.child;
-  }
-}
-
-class LocalizedMaterialApp extends StatefulWidget {
+class LocalizedMaterialApp extends StatelessWidget {
   const LocalizedMaterialApp(
       this.customOverlayWidget, this.customBetweenFabWidget,
       {Key key})
@@ -216,11 +168,6 @@ class LocalizedMaterialApp extends StatefulWidget {
   final LocaleWidgetBuilder customOverlayWidget;
   final WidgetBuilder customBetweenFabWidget;
 
-  @override
-  _LocalizedMaterialAppState createState() => _LocalizedMaterialAppState();
-}
-
-class _LocalizedMaterialAppState extends State<LocalizedMaterialApp> {
   @override
   Widget build(BuildContext context) {
     final routes = <String, WidgetBuilder>{
@@ -249,8 +196,8 @@ class _LocalizedMaterialAppState extends State<LocalizedMaterialApp> {
           supportedLocales: TrufiLocalization.supportedLocales,
           theme: context.watch<ThemeCubit>().state.activeTheme,
           home: HomePage(
-            customOverlayWidget: widget.customOverlayWidget,
-            customBetweenFabWidget: widget.customBetweenFabWidget,
+            customOverlayWidget: customOverlayWidget,
+            customBetweenFabWidget: customBetweenFabWidget,
           ),
         );
       },
